pyproject.toml
[tool.poetry]
name = "vallox_meter"
version = "0.1.0"
description = "Reads various metrics from a vallox KWL"
authors = ["ChristianSauer <christian.sauer@email.de>"]

[tool.poetry.dependencies]
python = "^3.7"
loguru = "^0.5.3"
pyyaml = "^5.3.1"
pyravendb = "^5.0.0"
typing-extensions = "^3.7.4"
vallox-websocket-api = "^2.6.0"

[tool.poetry.dev-dependencies]
jupyterlab = "^2.2.8"
flake8 = "^3.8.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
